s3/requirements.txt
python-dotenv
smart-open